+++ project-starter/react-app/src/components/NavBar/NavBar.css
@@ -1,0 +1,4 @@
+.nav-container {
+    background-color: #388186;
+    margin-top: 0;
+}
